refactor: lift WI-00015 phases into omotion/tuning.py, thin CLI
The phase engine (baseline/tune/crosscheck/finalize/diagnose-low) moves
into the package so the test-app's planned tune button (test-app#70) can
import and drive the same phases; scripts/wi15_runner.py becomes a thin
argparse front end. win32com import deferred so omotion.tuning imports
without pywin32 installed.

Refs #214
